Allow for nested FluxComponents
Sometimes, a store state getter needs to use a prop derived from the
state of another store. Previously, when using FluxComponent, this
required the creation of an extra composite component to receive the
props and pass them to another FluxComponent.

With this commit, props on a FluxComponent are automatically transferred
to children. This lets you nest FluxComponents, so that the inner state
getters have access to props retreived by outer state getters.