Return value to action caller
Previously, actions always returned undefined (or, in the case of async
actions, a promise that resolved to undefined). This was to enforce the
unidirectional data-flow model of Flux, and prevent users from using
actions as helper functions. This seemed to cause more trouble than its
worth, however, because there are legitimate reasons for a caller to use
the return value of an actions -- primarily, testing.

This PR reverses that decision. Now, the return value of an action
method is both sent through the dispatcher and returned to the caller.