Improved Store `register` methods
This cleans up Store's `register`, `registerAsync`, `registerAll`, and
`registerAllAsync` methods -- and the corresponding logic in
`Store#handler()` -- so that the code paths are simpler and easier to
follow. It also adds the ability for multiple handlers to be registered
for the same action.

This is also changes, in a breaking way, the args that are passed to
action handlers. Every action handler is now passed as the final
argument the entire payload from the dispatcher. This allows us to
provide extra information to the handlers; e.g. a dispatch id to
enable fine-grained optimistic updates (not implemented yet).

Async "begin" handlers no longer receive as arguments the same arguments
passed to the action -- they receive the dispatcher payload.